Segunda entrega sem autoexclusão + refactor packages

+++ src/main/java/com/sd/server/repositories/UserRepository.java
@@ -3,18 +3,14 @@ package com.sd.server.repositories;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.sd.server.DAO.JWTSessionDAO;
 import com.sd.server.DAO.UserDAO;
-import com.sd.server.auth.JwtTokenProvider;
-import com.sd.server.exceptions.EmailAlreadyUsedException;
-import com.sd.server.exceptions.NoSessionException;
-import com.sd.server.exceptions.UnauthorizedUserException;
-import com.sd.server.models.User;
-import com.sd.server.package_data.request.user.DeleteUserRequestData;
-import com.sd.server.package_data.request.user.GetUserRequestData;
-import com.sd.server.package_data.response.user.GetUserResponseData;
-import com.sd.server.packages.BasePackage;
-import com.sd.server.packages.user.CreateUserRequestPackage;
-import com.sd.server.packages.user.DeleteUserRequestPackage;
-import com.sd.server.packages.user.GetUserRequestPackage;
+import com.sd.server.Exceptions.EmailAlreadyUsedException;
+import com.sd.server.Exceptions.NoSessionException;
+import com.sd.server.Exceptions.UnauthorizedUserException;
+import com.sd.server.Models.User;
+import com.sd.server.Packages.data.request.user.*;
+import com.sd.server.Packages.data.response.user.FindUserResponseData;
+import com.sd.server.Packages.data.response.user.GetUserResponseData;
+import com.sd.server.Packages.BasePackage;
 import org.mindrot.jbcrypt.BCrypt;
 
 import java.util.List;
@@ -26,9 +22,9 @@ public class UserRepository {
     JWTSessionDAO jwtSessionDAO = new JWTSessionDAO();
 
     public BasePackage create(String action, String create_request) throws JsonProcessingException, UnauthorizedUserException, NoSessionException, EmailAlreadyUsedException {
-        CreateUserRequestPackage  request = CreateUserRequestPackage.fromJson(create_request, CreateUserRequestPackage.class);
-        int user_id = JwtTokenProvider.getUserId(request.getData().getToken());
-        this.validateUser(user_id);
+        BasePackage<CreateUserRequestData> request = BasePackage.fromJson(create_request, CreateUserRequestData.class);
+        int user_id = AuthRepository.getUserId(request.getData().getToken());
+        this.validateAdminUser(user_id);
         User user = new User(request.getData());
         user.setPassword(hashUserPassw(user.getPassword()));
         userDAO.addUserIfNotExistByEmail(user);
@@ -39,7 +35,7 @@ public class UserRepository {
         return BCrypt.hashpw(password, BCrypt.gensalt());
     }
 
-    private void validateUser(int user_id) throws UnauthorizedUserException, NoSessionException {
+    private void validateAdminUser(int user_id) throws UnauthorizedUserException, NoSessionException {
         if(!userDAO.isUserAdmin(user_id)){
             throw new UnauthorizedUserException();
         }
@@ -47,24 +43,64 @@ public class UserRepository {
             throw new NoSessionException();
         }
     }
+    private void validateLoggedUser(int user_id) throws UnauthorizedUserException, NoSessionException {
+        if(!jwtSessionDAO.hasUserActiveSession(user_id)){
+            throw new NoSessionException();
+        }
+    }
 
-    public BasePackage get(String action, String get_request) throws JsonProcessingException, NoSessionException, UnauthorizedUserException {
-        GetUserRequestPackage request = GetUserRequestPackage.fromJson(get_request, GetUserRequestPackage.class);
-        int user_id = JwtTokenProvider.getUserId(request.getData().getToken());
-        this.validateUser(user_id);
+    public BasePackage<FindUserResponseData> find(String action, String get_request) throws JsonProcessingException, NoSessionException, UnauthorizedUserException {
+        BasePackage<FindUserRequestData> request = BasePackage.fromJson(get_request, FindUserRequestData.class);
+        int logged_user_id = AuthRepository.getUserId(request.getData().getToken());
+        Long user_id = request.getData().getUser_id();
+        this.validateLoggedUser(logged_user_id);
+        User user = userDAO.getUserById(user_id);
+        FindUserResponseData response_data = new FindUserResponseData(user);
+        return new BasePackage<FindUserResponseData>(action,response_data,false,"Sucesso");
+    }
+
+    public BasePackage<FindUserResponseData> findSelf(String action, String get_request) throws JsonProcessingException, NoSessionException, UnauthorizedUserException {
+        BasePackage<FindSelfUserRequestData> request = BasePackage.fromJson(get_request, FindSelfUserRequestData.class);
+        int logged_user_id = AuthRepository.getUserId(request.getData().getToken());
+        this.validateLoggedUser(logged_user_id);
+        User user = userDAO.getUserById(logged_user_id);
+        FindUserResponseData response_data = new FindUserResponseData(user);
+        return new BasePackage<FindUserResponseData>(action,response_data,false,"Sucesso");
+    }
+
+    public BasePackage destroy(String action, String destroy_request) throws JsonProcessingException, NoSessionException, UnauthorizedUserException {
+        BasePackage<DeleteUserRequestData> request = BasePackage.fromJson(destroy_request, DeleteUserRequestData.class);
+        int adm_id = AuthRepository.getUserId(request.getData().getToken());
+        long user_id = request.getData().getUser_id();
+        this.validateAdminUser(adm_id);
+        userDAO.deleteUser(user_id);
+        jwtSessionDAO.deleteAllUserJWTSession(user_id);
+        return new BasePackage(action,false,"Usuário removido com sucesso");
+    }
+
+    public BasePackage<GetUserResponseData> get(String action, String get_request) throws JsonProcessingException, NoSessionException, UnauthorizedUserException {
+        BasePackage<GetUserRequestData> request = BasePackage.fromJson(get_request, GetUserRequestData.class);
+        int user_id = AuthRepository.getUserId(request.getData().getToken());
+        this.validateAdminUser(user_id);
         List<User> users = userDAO.getAllUsers();
         GetUserResponseData response_data = new GetUserResponseData(users);
         return new BasePackage(action,response_data,false,"Sucesso");
     }
-
-    public BasePackage destroy(String action, String get_request) throws JsonProcessingException, NoSessionException, UnauthorizedUserException {
-        DeleteUserRequestPackage request = DeleteUserRequestPackage.fromJson(get_request, DeleteUserRequestPackage.class);
-        int adm_id = JwtTokenProvider.getUserId(request.getData().getToken());
-        long user_id = request.getData().getUser_id();
-        this.validateUser(adm_id);
-        userDAO.deleteUser(user_id);
-        jwtSessionDAO.deleteAllUserJWTSession(user_id);
-        return new BasePackage(action,false,"Usuário removido com sucesso");
+    public BasePackage update(String action, String update_request) throws JsonProcessingException, NoSessionException, UnauthorizedUserException {
+        BasePackage<EditUserRequestData> request = BasePackage.fromJson(update_request, EditUserRequestData.class);
+        int user_id = AuthRepository.getUserId(request.getData().getToken());
+        this.validateLoggedUser(user_id);
+        User user = userDAO.getUserById(request.getData().getId());
+        if(AuthRepository.isAdmin(request.getData().getToken())){
+            user.setType(request.getData().getType());
+        }
+        if(request.getData().getPassword() != null){
+            user.setPassword(hashUserPassw(request.getData().getPassword()));
+        }
+        user.setName(request.getData().getName());
+        user.setEmail(request.getData().getEmail());
+        userDAO.updateUser(user);
+        return new BasePackage(action,null,false,"Sucesso");
     }
 }
